feat: getLayout削除
