+++ client/src/pages/mypage.tsx
@@ -1,10 +1,8 @@
-import type { ReactElement } from 'react';
+import type { NextPage } from 'next';
 import { Text, Spinner } from '@chakra-ui/react';
 import useGetSignInUser from 'src/hooks/useGetSignInUser';
-import SignedInLayout from 'src/components/layouts/SignedInLayout';
-import type NextPageWithLayout from 'src/types/nextPageWithLayout';
 
-const Mypage: NextPageWithLayout = () => {
+const Mypage: NextPage = () => {
   const signInUser = useGetSignInUser();
 
   if (!signInUser) return <Spinner />;
@@ -18,9 +16,5 @@ const Mypage: NextPageWithLayout = () => {
     </>
   );
 };
-
-Mypage.getLayout = (page: ReactElement) => (
-  <SignedInLayout>{page}</SignedInLayout>
-);
 
 export default Mypage;
