+++ client/src/pages/sign_up.tsx
@@ -1,4 +1,4 @@
-import { ReactElement } from 'react';
+import type { NextPage } from 'next';
 import { useForm } from 'react-hook-form';
 import { Box, Text } from '@chakra-ui/react';
 import { zodResolver } from '@hookform/resolvers/zod';
@@ -6,10 +6,8 @@ import InputForm from 'src/components/InputForm';
 import OutlineButton from 'src/components/OutlineButton';
 import { SignUpSchema, SignUpForm } from 'src/formSchemas/signUpSchema';
 import useSignUp from 'src/hooks/useSignUp';
-import UnSignedInLayout from 'src/components/layouts/UnSignedInLayout';
-import type NextPageWithLayout from 'src/types/nextPageWithLayout';
 
-const SignUp: NextPageWithLayout = () => {
+const SignUp: NextPage = () => {
   const signUp = useSignUp();
   const {
     handleSubmit,
@@ -71,9 +69,5 @@ const SignUp: NextPageWithLayout = () => {
     </>
   );
 };
-
-SignUp.getLayout = (page: ReactElement) => (
-  <UnSignedInLayout>{page}</UnSignedInLayout>
-);
 
 export default SignUp;
