--- client/src/types/nextPageWithLayout.ts
@@ -1,8 +1,0 @@
-import { ReactElement, ReactNode } from 'react';
-import type { NextPage } from 'next';
-
-type NextPageWithLayout = NextPage & {
-  getLayout?: (page: ReactElement) => ReactNode;
-};
-
-export default NextPageWithLayout;
